backend(OLD_PROJECT)/requirements.txt
# Core requirements
Django==4.2.13
djangorestframework==3.14.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
# Authentication and security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-otp==1.3.0  # For two-factor authentication
pyotp==2.9.0  # For TOTP-based authentication
argon2-cffi==23.1.0  # For secure password hashing
django-permissions-policy==4.18.0  # Security headers

# Data validation and parsing
pydantic==2.6.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
lxml==4.9.3
python-magic==0.4.27  # For file type detection

# Database and caching
django-filter==23.5
django-redis==5.4.0
redis==5.0.1
django-cacheops==7.0.2

# API documentation
drf-spectacular==0.27.0
drf-yasg==1.21.7

# Text processing and search
nltk==3.8.1
django-haystack==3.2.1
Pillow==10.1.0  # For image processing
pgvector==0.2.4  # For PostgreSQL vector operations

# Rate limiting and throttling
django-ratelimit==4.1.0
django-defender==0.9.7  # For login throttling

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Monitoring and logging
sentry-sdk==1.39.1
django-structlog==6.0.1

# Utilities
django-storages==1.14.2  # For file storage backends
djangorestframework-camel-case==1.4.2  # For converting between snake_case and camelCase
django-cleanup==8.0.0  # For cleaning up files